Combine EVM wallet modules
