+++ index.html
@@ -278,7 +278,7 @@
         <div class="modal-content">
           <ul class="menu-list">
             <li class="menu-item" id="openInvite" data-icon="mail">Invite</li>
-            <li class="menu-item" id="openAssets" data-icon="assets">Assets</li>
+            <li class="menu-item" id="openAssets" data-icon="assets">EVM Assets</li>
             <li class="menu-item" id="openFarm" data-icon="trending-up">Farm</li>
             <li class="menu-item" id="openValidator" data-icon="shield">Validator</li>
             <li class="menu-item" id="openDao" data-icon="dao" style="display: none;">DAO</li>
